WW-5459 Move new Action class into action package
