+++ core/src/main/java/org/apache/struts2/ActionInvocation.java
@@ -19,6 +19,7 @@
 package org.apache.struts2;
 
 import com.opensymphony.xwork2.ActionChainResult;
+import org.apache.struts2.action.Action;
 import org.apache.struts2.interceptor.PreResultListener;
 import org.apache.struts2.result.Result;
 import org.apache.struts2.util.ValueStack;
