+++ core/src/main/java/org/apache/struts2/ActionSupport.java
@@ -29,6 +29,7 @@ import com.opensymphony.xwork2.inject.Inject;
 import com.opensymphony.xwork2.util.ValueStack;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.apache.struts2.action.Action;
 import org.apache.struts2.interceptor.ValidationAware;
 
 import java.io.Serializable;
