+++ core/src/test/java/com/opensymphony/xwork2/config/providers/ConfigurationProviderOgnlAllowlistTest.java
@@ -67,7 +67,7 @@ public class ConfigurationProviderOgnlAllowlistTest extends XWorkJUnit4TestCase 
                 Class.forName("org.apache.struts2.interceptor.Interceptor"),
                 Class.forName("org.apache.struts2.interceptor.ConditionalInterceptor"),
                 Class.forName("org.apache.struts2.result.Result"),
-                Class.forName("org.apache.struts2.Action"),
+                Class.forName("org.apache.struts2.action.Action"),
                 Class.forName("org.apache.struts2.Validateable"),
                 Class.forName("org.apache.struts2.interceptor.ValidationAware")
         );
@@ -99,7 +99,7 @@ public class ConfigurationProviderOgnlAllowlistTest extends XWorkJUnit4TestCase 
                 Class.forName("org.apache.struts2.interceptor.Interceptor"),
                 Class.forName("org.apache.struts2.interceptor.ConditionalInterceptor"),
                 Class.forName("org.apache.struts2.result.Result"),
-                Class.forName("org.apache.struts2.Action"),
+                Class.forName("org.apache.struts2.action.Action"),
                 Class.forName("org.apache.struts2.Validateable"),
                 Class.forName("org.apache.struts2.interceptor.ValidationAware")
         );
@@ -130,7 +130,7 @@ public class ConfigurationProviderOgnlAllowlistTest extends XWorkJUnit4TestCase 
                 Class.forName("org.apache.struts2.interceptor.Interceptor"),
                 Class.forName("org.apache.struts2.interceptor.ConditionalInterceptor"),
                 Class.forName("org.apache.struts2.result.Result"),
-                Class.forName("org.apache.struts2.Action"),
+                Class.forName("org.apache.struts2.action.Action"),
                 Class.forName("org.apache.struts2.Validateable"),
                 Class.forName("org.apache.struts2.interceptor.ValidationAware")
         );
